move argparse to utils
